Move certificate validation into a module
We need to be able to use this in a couple of different places and not
always directly linked to an event model.
